add step deploy in git actions v.114

+++ backend/foodgram/api/views.py
@@ -185,7 +185,7 @@ class FavoriteViewSet(CreateDestroyMixin):
         if not Favorite.objects.filter(user=user, recipe=recipe):
             Favorite.objects.create(user=user, recipe=recipe)
             queryset = get_object_or_404(Recipe, pk=pk)
-            serializer = FavoritRecipeSerializer(queryset)
+            serializer = FavoriteRecipeSerializer(queryset)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
         return Response(
@@ -196,9 +196,14 @@ class FavoriteViewSet(CreateDestroyMixin):
     def delete(self, request, pk):
         user = request.user
         recipe = get_object_or_404(Recipe, pk=pk)
+        if not Favorite.objects.filter(user=user, recipe=recipe):
+            return Response(
+                f'Рецепт {recipe.name} не добавлен в избранное',
+                status=status.HTTP_400_BAD_REQUEST
+            )
         Favorite.objects.filter(user=user, recipe=recipe).delete()
         return Response(
-            f'Рецепт {recipe.name} удалён из избранного',
+            f'Рецепт {recipe.name} успешно удалён из избранного',
             status=status.HTTP_204_NO_CONTENT
         )
